rename ThemeProvider to themeConfig without a default export
