+++ web/src/app/index.tsx
@@ -7,7 +7,7 @@ import { Provider as ReduxProvider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 import { ApolloProvider } from '@apollo/client'
 import { StyledEngineProvider } from '@mui/material/styles'
-import ThemeProvider from './main/components/ThemeProvider'
+import { ThemeProvider } from './main/themeConfig'
 import { GraphQLClient } from './apollo'
 import './styles'
 import App from './main/App'
